refactor(gutenberg): set form select slate as global component
